separated function into routes and controller
