+++ backend/app.js
@@ -6,10 +6,7 @@ import bodyParser from 'body-parser'
 import userRoutes from './routes/user.js'
 import postRoutes from './routes/post.js'
 import profileRoutes from './routes/profile.js'
-import { multerUploads } from './utils/multer.js'
-import { cloudinaryConfig, uploader } from './utils/cloudinary.js'
-import path from 'path'
-import DatauriParser from 'datauri/parser.js'
+import { cloudinaryConfig } from './utils/cloudinary.js'
 
 
 const app = express()
@@ -36,34 +33,6 @@ app.get('/', (req, res) => {
     res.send('Hello World!')
   })
 
-  // Fix this
-app.post('/testupload', multerUploads.single('image'), (req, res) => { 
-
-const parser = new DatauriParser()
-
-const extName = path.extname(req.file.originalname).toString();
-const file64 = parser.format(extName, req.file.buffer);
-
-
-    // const file = dataUri(req)
-    // console.log(file)
-    return uploader.upload(file64.content).then((result) => {
-        const image = result.url
-        return res.status(200).json({
-            message: 'Your image has been uploaded to cloudinary',
-            data: {
-                image
-            }
-        }).then()
-        
-
-    }).catch((err) => res.status(400).json({
-        message: 'Something went wrong with your upload to cloudinary',
-        data: { err }
-    }))
-
-})
-  
 
 // Backend API routes
 app.use('/api/posts', postRoutes)
